refactor: centralize constants definition
